feat: added os and hash operations & reorganized code
