--- src/utils/commands.js
@@ -1,9 +1,0 @@
-export { up } from './up.js';
-export { ls } from './ls.js';
-export { cd } from './cd.js';
-export { cat } from './cat.js';
-export { add } from './add.js';
-export { rn } from './rn.js';
-export { cp } from './cp.js';
-export { mv } from './mv.js';
-export { rm } from './rm.js';
