+++ src/utils/operations.js
@@ -1,0 +1,4 @@
+export { up, cd, ls } from './nwd.js';
+export { cat, add, rn, cp, rm, mv } from './fs.js';
+export { os } from './os.js';
+export { hash } from './hash.js';
